test: Move toMatchTaskWithDescription() to separate file
